break out notebook as separate package
