+++ packages/_meta/src/index.ts
@@ -1,2 +1,3 @@
 import '@deathbeds/jupyterlab-drawio';
 import '@deathbeds/jupyterlab-drawio-pdf';
+import '@deathbeds/jupyterlab-drawio-notebook';
