+++ packages/jupyterlab-drawio-notebook/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@deathbeds/jupyterlab-drawio-notebook",
+  "version": "0.8.0-alpha0",
+  "dependencies": {
+    "@deathbeds/jupyterlab-drawio": "0.8.0-alpha0",
+    "@jupyterlab/notebook": "^2.0.0"
+  },
+  "files": [
+    "{lib,schema,style,src}/**/*.{ts,tsx,eot,gif,html,jpg,js,js.map,json,png,svg,woff2,css}",
+    "LICENSE.txt"
+  ],
+  "jupyterlab": {
+    "extension": "lib/plugin.js"
+  },
+  "main": "lib/index.js"
+}
