+++ packages/jupyterlab-drawio-notebook/src/index.ts
@@ -1,0 +1,2 @@
+export const NS = '@deathbeds/jupyterlab-drawio-notebook';
+export const PLUGIN_ID = `${NS}:plugin`;
